ADA/Practica_JDBC_1/src/main/java/datos/ClienteDAO.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package datos;

import static datos.Conexion.getConnection;
import java.util.*;
import java.sql.*;
import domain.Cliente;

/**
 *
 * @author juasanlop2
 */
public class ClienteDAO {

    private static final String SQL_SELECT = "SELECT id_cliente, saldo, limite_credito, descuento FROM cliente";
    private static final String SQL_INSERT = "INSERT INTO cliente (saldo, limite_credito, descuento) VALUES (?,?,?)";
    private static final String SQL_UPDATE = "UPDATE cliente SET saldo = ?, limite_credito = ?, descuento = ? WHERE id_cliente = ?";
    private static final String SQL_DELETE = "DELETE FROM cliente WHERE id_cliente = ?";

    public List<Cliente> seleccionar() throws SQLException {
        Connection conn = null;
        PreparedStatement stmt = null;
        ResultSet rs = null;
        Cliente cliente = null;
        List<Cliente> clientes = new ArrayList<>();

        try {
            conn = getConnection();
            stmt = conn.prepareStatement(SQL_SELECT);
            rs = stmt.executeQuery();

            while (rs.next()) {
                int idCliente = rs.getInt("id_cliente");
                float saldo = rs.getFloat("saldo");
                float limiteCredito = rs.getFloat("limite_credito");
                float descuento = rs.getFloat("descuento");
                cliente = new Cliente(idCliente, saldo, limiteCredito, descuento);
                clientes.add(cliente);
            }

        } catch (SQLException ex) {
            ex.printStackTrace(System.out);
        } finally {
            Conexion.close(conn);
            Conexion.close(rs);
            Conexion.close(stmt);
        }

        return clientes;
    }

    public int insertar(Cliente cliente) {
        Connection conn = null;
        PreparedStatement stmt = null;

        int registros = 0;

        try {
            conn = Conexion.getConnection();
            stmt = conn.prepareStatement(SQL_INSERT);
            stmt.setFloat(1, cliente.getSaldo());
            stmt.setFloat(2, cliente.getLimiteCredito());
            stmt.setFloat(3, cliente.getDescuento());
            registros = stmt.executeUpdate();

        } catch (SQLException ex) {
            ex.printStackTrace(System.out);
        } finally {
            try {
                Conexion.close(stmt);
                Conexion.close(conn);

            } catch (SQLException ex) {
                ex.printStackTrace(System.out);
            }
        }
        return registros;
    }

    public int update(Cliente clienteID, Cliente cliente) {
        Connection conn = null;
        PreparedStatement stmt = null;

        int registros = 0;

        try {
            conn = Conexion.getConnection();
            stmt = conn.prepareStatement(SQL_UPDATE);
            stmt.setFloat(1, cliente.getSaldo());
            stmt.setFloat(2, cliente.getLimiteCredito());
            stmt.setFloat(3, cliente.getDescuento());
            stmt.setInt(4, clienteID.getIdCliente());
            registros = stmt.executeUpdate();

        } catch (SQLException ex) {
            ex.printStackTrace(System.out);
        } finally {
            try {
                Conexion.close(stmt);
                Conexion.close(conn);

            } catch (SQLException ex) {
                ex.printStackTrace(System.out);
            }
        }
        return registros;
    }

    public int delete(Cliente clienteID) {
        Connection conn = null;
        PreparedStatement stmt = null;

        int registros = 0;

        try {
            conn = Conexion.getConnection();
            stmt = conn.prepareStatement(SQL_DELETE);
            stmt.setInt(1, clienteID.getIdCliente());
            registros = stmt.executeUpdate();

        } catch (SQLException ex) {
            ex.printStackTrace(System.out);
        } finally {
            try {
                Conexion.close(stmt);
                Conexion.close(conn);

            } catch (SQLException ex) {
                ex.printStackTrace(System.out);
            }
        }
        return registros;
    }
}
